Changed TinyImageDataset -> SectionImage and moved it into the components folder so they can be used universally
