+++ frontend/src/components/SectionImage.vue
